refactor : 랭킹 API 인터페이스 삭제, RankingServiceImpl 을 RankingService 로 변경
